+++ src/IPFees.Calculator/Evaluator/NodeFunctionCall.cs
@@ -17,10 +17,10 @@ namespace IPFees.Evaluator
         string _functionName;
         Node[] _arguments;
 
-        public override double Eval(IContext ctx)
+        public override decimal Eval(IContext ctx)
         {
             // Evaluate all arguments
-            var argVals = new double[_arguments.Length];
+            var argVals = new decimal[_arguments.Length];
             for (int i=0; i<_arguments.Length; i++)
             {
                 argVals[i] = _arguments[i].Eval(ctx);
