+++ src/IPFees.Calculator/Evaluator/NodeNumber.cs
@@ -3,14 +3,14 @@
     // NodeNumber represents a literal number in the expression
     class NodeNumber : Node
     {
-        public NodeNumber(double number)
+        public NodeNumber(decimal number)
         {
             _number = number;
         }
 
-        double _number;             // The number
+        decimal _number;             // The number
 
-        public override double Eval(IContext ctx)
+        public override decimal Eval(IContext ctx)
         {
             // Just return it.  Too easy.
             return _number;
