+++ src/IPFees.Calculator/Evaluator/NodeUnary.cs
@@ -7,16 +7,16 @@ namespace IPFees.Evaluator
     {
         // Constructor accepts the two nodes to be operated on and function
         // that performs the actual operation
-        public NodeUnary(Node rhs, Func<double, double> op)
+        public NodeUnary(Node rhs, Func<decimal, decimal> op)
         {
             _rhs = rhs;
             _op = op;
         }
 
         Node _rhs;                              // Right hand side of the operation
-        Func<double, double> _op;               // The callback operator
+        Func<decimal, decimal> _op;               // The callback operator
 
-        public override double Eval(IContext ctx)
+        public override decimal Eval(IContext ctx)
         {
             // Evaluate RHS
             var rhsVal = _rhs.Eval(ctx);
